+++ src/components/ServerModal.vue
@@ -120,6 +120,7 @@
 import { useVuelidate } from '@vuelidate/core'
 import { required } from '@vuelidate/validators'
 import { defineComponent } from 'vue'
+
 import { ServerConfig } from '../interfaces'
 import store from '../store'
 import ServerConfigs from './ServerConfigs.vue'
