+++ src/pages/Server/ServerLogs.vue
@@ -18,7 +18,7 @@
   /> -->
 </template>
 
-<script lang="ts">
+<script lang='ts'>
 import { defineComponent } from 'vue'
 
 export default defineComponent({
